requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
claude-monitor==3.1.0
click==8.3.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
executing==2.2.1
fastapi==0.121.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
langchain-core==1.0.4
langchain-openai==1.0.2
langsmith==0.4.42
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
narwhals==2.11.0
nest-asyncio==1.6.0
numpy==2.2.6
openai==2.7.2
opencv-python-headless==4.12.0.88
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.2
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pydeck==0.9.1
pydicom==3.0.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.49.3
streamlit==1.51.0
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
wcwidth==0.2.14
zstandard==0.25.0
